@@ -16,7 +16,7 @@ export { PayeeForm } from './PayeeForm'
 export { DeleteWallet } from './DeleteWallet'
 export { getPasswordConfirmation } from './PasswordConfirmation'
 export { TransferForm } from './TransferForm'
-export { TransactionOperation } from './TransactionOperation'
+export { TransactionCard } from './TransactionOperation'
 export { Transactions } from './Transactions'
 export { ConnectionSelector } from './ConnectionSelector'
 export { ConnectionSettings } from './ConnectionSettings'
@@ -24,3 +24,4 @@ export { ConnectionForm } from './ConnectionForm'
 export { Transfer } from './Transfer'
 export { AccountPageSelector } from './AccountPageSelector'
 export { Sign } from './Sign'
+export { SignTransactionForm } from './SignTransactionForm'
