@@ -4,30 +4,32 @@ import { flatten, get } from 'lodash'
 import { Connection, TransactionOperationView } from '@types'
 const STROOPS_IN_ONE_KINESIS = 10000000
 
-export const operationErrorCodeToMessage = {
-  op_low_reserve: 'Transfer amount is lower than the base reserve',
-  op_already_exist: 'The account already exists.',
-  op_no_destination: 'The target payment account does not exist.',
-  unknown_error: 'An error occured with the transaction.',
+export enum OperationErrors {
+  op_low_reserve = 'Transfer amount is lower than the base reserve',
+  op_already_exist = 'The account already exists.',
+  op_no_destination = 'The target payment account does not exist.',
+  op_bad_auth = 'Invalid transaction signature',
+}
+
+export enum TransactionErrors {
+  tx_insufficient_fee = 'The fee on the transaction was too low',
+  tx_insufficient_balance = 'Insufficient account balance',
+  tx_bad_auth = 'Invalid transaction signature',
+  tx_bad_auth_extra = 'Too many signatures provided',
+  tx_bad_seq = 'Invalid account sequence',
 }
 
 export function transactionErrorCodeToMessage(txError: string, opError: string) {
-  const errorToMessage = {
-    tx_insufficient_fee: 'The fee on the transaction was too low',
-    tx_insufficient_balance: 'Insufficient account balance',
-    tx_bad_auth: 'Invalid transaction signature',
-    tx_bad_auth_extra: 'Too many signatures provided',
-    tx_bad_seq: 'Invalid account sequence',
-    tx_failed: operationErrorCodeToMessage[opError],
-    unknown_error: 'An error occured with the transaction.',
-  }
-
-  return errorToMessage[txError]
+  return (
+    OperationErrors[opError] ||
+    TransactionErrors[txError] ||
+    'An error occurred with the transaction'
+  )
 }
 
 export function getTransactionErrorMessage(failedTxPayload: any) {
-  const txErrorCode = get(failedTxPayload, 'data.extras.result_codes.transaction', 'unknown_error')
-  const opErrorCode = get(failedTxPayload, 'data.extras.result_codes.operations[0]', 'unknown_error')
+  const txErrorCode = get(failedTxPayload, 'data.extras.result_codes.transaction')
+  const opErrorCode = get(failedTxPayload, 'data.extras.result_codes.operations[0]')
   return transactionErrorCodeToMessage(txErrorCode, opErrorCode)
 }
 
@@ -36,18 +38,19 @@ export function getServer(connection: Connection): Server {
   return new Server(connection.horizonURL)
 }
 
-export async function getFeeInStroops(
-  server: Server,
-  amountInKinesis: number,
-): Promise<string> {
-  const mostRecentLedger = await server.ledgers().order('desc').call()
+export async function getFeeInStroops(server: Server, amountInKinesis: number): Promise<string> {
+  const mostRecentLedger = await server
+    .ledgers()
+    .order('desc')
+    .call()
   const {
     base_percentage_fee: basePercentageFee,
     base_fee_in_stroops: baseFeeInStroops,
   } = mostRecentLedger.records[0]
   const basisPointsToPercent = 10000
 
-  const percentageFee = Number(amountInKinesis) * basePercentageFee / basisPointsToPercent * STROOPS_IN_ONE_KINESIS
+  const percentageFee =
+    ((Number(amountInKinesis) * basePercentageFee) / basisPointsToPercent) * STROOPS_IN_ONE_KINESIS
 
   return String(Math.ceil(percentageFee + baseFeeInStroops))
 }
@@ -66,8 +69,14 @@ export async function getTransactions(
 ): Promise<TransactionOperationView[]> {
   const server = getServer(connection)
   try {
-    const transactionPage = await server.transactions().forAccount(accountKey).order('desc').call()
-    const nestedArray = await Promise.all(transactionPage.records.map((t) => transactionWithOperations(t, accountKey)))
+    const transactionPage = await server
+      .transactions()
+      .forAccount(accountKey)
+      .order('desc')
+      .call()
+    const nestedArray = await Promise.all(
+      transactionPage.records.map(t => transactionWithOperations(t, accountKey)),
+    )
     return flatten(nestedArray)
   } catch (e) {
     return []
@@ -79,12 +88,14 @@ async function transactionWithOperations(
   accountKey: string,
 ): Promise<TransactionOperationView[]> {
   const operationsPage = await transaction.operations()
-  return operationsPage.records.map((operation): TransactionOperationView => ({
-    operation,
-    date: new Date(transaction.created_at),
-    fee: (Number(transaction.fee_paid) / STROOPS_IN_ONE_KINESIS).toFixed(5) ,
-    isIncoming: transaction.source_account === accountKey,
-    memo: transaction.memo,
-    source: transaction.source_account,
-  }))
+  return operationsPage.records.map(
+    (operation): TransactionOperationView => ({
+      operation,
+      date: new Date(transaction.created_at),
+      fee: (Number(transaction.fee_paid) / STROOPS_IN_ONE_KINESIS).toFixed(5),
+      isIncoming: transaction.source_account === accountKey,
+      memo: transaction.memo,
+      source: transaction.source_account,
+    }),
+  )
 }
