@@ -1,4 +1,11 @@
-import { addPayee, addWallet, changeTransferView, changeWalletView, deleteWallet, selectWallet } from '@actions'
+import {
+  addPayee,
+  addWallet,
+  changeTransferView,
+  changeWalletView,
+  deleteWallet,
+  selectWallet,
+} from '@actions'
 import { RootAction } from '@store'
 import { TransferView, WalletView } from '@types'
 import { combineReducers } from 'redux'
@@ -12,24 +19,29 @@ export interface ViewState {
 export const view = combineReducers<ViewState, RootAction>({
   walletView: (state = WalletView.create, action) => {
     switch (action.type) {
-      case getType(changeWalletView): return action.payload
+      case getType(changeWalletView):
+        return action.payload
 
       case getType(selectWallet):
       case getType(addWallet):
         return WalletView.dashboard
 
-      case getType(deleteWallet): return WalletView.create
-      default: return state
+      case getType(deleteWallet):
+        return WalletView.create
+      default:
+        return state
     }
   },
   transferView: (state = TransferView.transfer, action) => {
     switch (action.type) {
-      case getType(changeTransferView): return action.payload
+      case getType(changeTransferView):
+        return action.payload
 
       case getType(addPayee):
       case getType(changeWalletView):
         return TransferView.transfer
-      default: return state
+      default:
+        return state
     }
   },
 })
