@@ -1,5 +1,5 @@
+import * as actions from '@actions'
 import { ActionsUnion } from 'typesafe-actions'
-import * as actions from './actions'
 
 export type RootAction = ActionsUnion<typeof actions>
 
